requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv
requests
pre-commit
